% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggir_swarm.R
\name{clean_user_dir}
\alias{clean_user_dir}
\title{Empty the biowR cache}
\usage{
clean_user_dir()
}
\value{
nothing returned
}
\description{
removes files from the tools::R_user_dir("biowR","cache") directory
}
\seealso{
\code{\link[tools:userdir]{tools::R_user_dir()}}
}
