% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{concat_mangled_files}
\alias{concat_mangled_files}
\title{Concatenates files that would have the name if they were not mangled}
\usage{
concat_mangled_files(fileDir, unmangled)
}
\arguments{
\item{fileDir}{directory containing all the mangled files}

\item{unmangled}{the unmangled filename.}
}
\description{
Takes the unmangled name and finds all files that would be mangled to that file and
concatentes them all together.
}
