% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggir_swarm.R
\name{ensure_absolute_path}
\alias{ensure_absolute_path}
\title{Ensure that the Path passed in is an absolute path}
\usage{
ensure_absolute_path(path_in)
}
\arguments{
\item{path_in}{a possibly relative path}
}
\value{
an absolute path
}
\description{
Ensure that the Path passed in is an absolute path
}
