% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggir_swarm.R
\name{getStartAndEndJobs}
\alias{getStartAndEndJobs}
\title{calculates the number of files to run on a hyperthread
n_jobs (integer)
n_cpu (integer)
returns a list containing a vector of integers startIndex of jobs
and a list of the endIndex of a job.  This is INCLUSIVE.}
\usage{
getStartAndEndJobs(f0, f1, n_core, ht)
}
\arguments{
\item{f0}{\itemize{
\item first index run
}}

\item{f1}{\itemize{
\item last index run
}}

\item{n_core}{\itemize{
\item number of cores requested
}}

\item{ht}{\itemize{
\item true/false use hyperthreading?
}}
}
\description{
calculates the number of files to run on a hyperthread
n_jobs (integer)
n_cpu (integer)
returns a list containing a vector of integers startIndex of jobs
and a list of the endIndex of a job.  This is INCLUSIVE.
}
