% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{mangle_filename}
\alias{mangle_filename}
\alias{demangle_filename}
\title{Mangles the filename to prevent name collision}
\usage{
mangle_filename(x)

demangle_filename(x)
}
\arguments{
\item{x}{A filename containing a file extension of at least 1 character.
The argument x can be a vector of filenames. For mangle_filename, although each value
must be unique or else you could have two files with the same name.
Best practice is to make x simply a string.  For demangle_filename, using a vector
causes no problems}
}
\value{
the vector of mangeled or demangled filenames
}
\description{
The file names are mangled by insert the date-time right before the file extension.
Because of this, all file names are required to have at least a one letter extension.
}
\examples{
# create a vector of files...
files <- c("~/file1.txt","~/file2.txt")
mangled <- mangle_filename(files)
demangled <- demangle_filename(mangled)
# check that the demangle files are the same as the original files.
all(demangled == files)
}
