% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggir_swarm.R, R/ggir_functions.R
\name{run GGIR}
\alias{run GGIR}
\alias{run_stage_one}
\alias{run_stages_2_5}
\title{Run GGIR on the NIH biowulf.}
\usage{
run_stage_one(cwa_root, results_root, json_args, f0, f1)

run_stages_2_5(part1_output_dir, json_args = "", f0, f1)
}
\arguments{
\item{cwa_root}{where the CWA files are stored.}

\item{results_root}{where to write the results}

\item{json_args}{a json file containing a list of arguments passed into ggir}

\item{f0}{the starting index of the files in cwa_root}

\item{f1}{the ending index of the files in cwa_root}

\item{part1_output_dir}{the \strong{output} directory from running part 1 of the GGIR}
}
\description{
On the NIH biowulf, reading files directly from the /data disk is EXTREMELY slow.  So
we stage the data, files [f0..f1)  on local disk and run all the local files.

The goal is to make a pipeline that is swarmable.  Nothing is returned by
this function, however, files are written in the results root.

The output of run_stage_one should would with the input of run_stage_2_5.
}
\details{
The \strong{output} directory used in parts 2-5 is different from the results root
from part 1.  The output directory is inside the results root and has a directory
path of \emph{\strong{RESULTS_ROOT/output_<studyname>}}
}
