% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenames.R
\name{stage_files}
\alias{stage_files}
\title{Stage files from one dir to another.}
\usage{
stage_files(fromDir, toDir, mangle = FALSE, pattern, f0 = 0, f1)
}
\arguments{
\item{fromDir}{directory you are copying data from}

\item{toDir}{directory where the data is going to}

\item{mangle}{should the filenames be mangled while staging the files?}

\item{pattern}{(optional) pattern used to select files to stage see \code{\link[=list.files]{list.files()}}}

\item{f0}{(optional) the starting file index, if you only want to copy a
subset of the file (e.g. from file 3 to file 7).  By default is 0.}

\item{f1}{(optional) the ending file index, if you only want to copy a
subset of the file.}
}
\description{
On the biowulf system, the data disk is very slow for constant IO.  You
will want to stage the data from the data disk to the local scratch disk.
}
