% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggir_functions.R
\name{unstage_all}
\alias{unstage_all}
\title{Copy results from the staging back to the results directory}
\usage{
unstage_all(stage_output, results_output)
}
\arguments{
\item{stage_output}{the output directory on the stage disk}

\item{results_output}{the central output director}
}
\value{
a tibble with columns (file,name,relpath,mangle,ok)\tabular{lllll}{
   \strong{file} \tab \strong{name} \tab \strong{relpath} \tab \strong{mangle} \tab \strong{ok} \cr
   relative path of the files \tab basename of the file. \tab relative name of the files' directory \tab Was the file mangled? \tab Did the file copies ok? \cr
}
}
\description{
The way we run GGIR steps 2-5 on biowulf is to copy a set of results
from step 1 to a local staging scratch directory on one of the nodes.
The scratch directory is only available on the compute node.
Then steps 2-5 are run.

This function takes the results from steps 2-5 and returns it to the
central results directory, which is a networked disk.

The output directory is named \strong{output_studyname}.
}
