% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggir_swarm.R
\name{write_R_script}
\alias{write_R_script}
\alias{write_stage1_R_script}
\alias{write_stage2_5_R_script}
\title{Write the R script for running GGIR}
\usage{
write_stage1_R_script(script_dir = tools::R_user_dir("biowR", "cache"))

write_stage2_5_R_script(script_dir = tools::R_user_dir("biowR", "cache"))
}
\arguments{
\item{script_dir}{the directory where the R script is written}
}
\value{
the name of the script file (invisibly)
}
\description{
write_stage1_R_script and write_stage2_5_R_script are called from
write_stage1_swarmfile or write_stage2_5_swarmfile respectively.
}
\seealso{
\link{write_stage1_swarmfile}
}
