% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggir_swarm.R
\name{write_stage1_R_script}
\alias{write_stage1_R_script}
\title{Write the R script for stage1 of GGIR}
\usage{
write_stage1_R_script(scriptDir = tools::R_user_dir("biowR"))
}
\arguments{
\item{scriptDir}{the directory where the R script is written}
}
\value{
the name of the script file (invisiblly)
}
\description{
This is called from write_stage1_swarmfile
}
\seealso{
\code{\link[=write_stage1_swarmfile]{write_stage1_swarmfile()}}
}
