% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggir_swarm.R
\name{write_stage1_swarmfile}
\alias{write_stage1_swarmfile}
\alias{write_stage2_5_swarmfile}
\alias{rewrite_stage1_swarmfile}
\alias{rewrite_stage2_5_swarmfile}
\title{Prepares the swarmfile/script file for Stage 1 of the GGIR analysis}
\usage{
write_stage1_swarmfile(
  script_dir = tools::R_user_dir("biowR", "cache"),
  cwa_root,
  results_root,
  json_args = "",
  f0,
  f1,
  n_core,
  ht = FALSE
)

write_stage2_5_swarmfile(
  script_dir = tools::R_user_dir("biowR", "cache"),
  output_dir,
  json_args,
  f0,
  f1,
  n_core,
  ht = FALSE
)

rewrite_stage1_swarmfile(
  script_dir = tools::R_user_dir("biowR", "cache"),
  cwa_root,
  results_root,
  json_args = "",
  indices,
  files,
  ht = FALSE
)

rewrite_stage2_5_swarmfile(
  script_dir = tools::R_user_dir("biowR", "cache"),
  output_dir,
  json_args,
  indices,
  ht = FALSE
)
}
\arguments{
\item{script_dir}{The directory where the scripts are written.}

\item{cwa_root}{The directory where the accelerometer files are stored.}

\item{results_root}{The root for the results directory.}

\item{json_args}{An optional json file with parameters for GGIR}

\item{f0}{The start index}

\item{f1}{The end index.  The file with index f1 is run.}

\item{n_core}{The number of jobs you would like to swarm at once.}

\item{ht}{If you want to use hyper threading, set ht to TRUE.}

\item{output_dir}{The output_<study$gt; directory in the result_root}

\item{indices}{A vector of file indices to (re)analyze}

\item{files}{The files names to (re)analyze.
directory from stage 1}
}
\value{
invisibly returns the name of the swarmfile
}
\description{
Pay close attention to the parameters passed in. Currently, the code assumes that
each job takes 120 minutes to run.  This is to ensure that enough time is allocate to
run the job.
}
\details{
Because CRAN policy does not allow the scripts to be placed in the user's home directory without
permission, the default location of the scripts is in tools::R_user_dir("biowR","cache"), which on biowulf
is ~/.cache/R/biowR.  I would suggest setting the script_dir argument to something useful.
}
