% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggir_swarm.R
\name{write_stage1_swarmfile}
\alias{write_stage1_swarmfile}
\alias{write_stage2_5_swarmfile}
\alias{write_R_script}
\alias{write_stage1_R_script}
\alias{write_stage2_5_R_script}
\title{Prepares the swarmfile/script file for Stage 1 of the GGIR analysis}
\usage{
write_stage1_swarmfile(
  scriptDir = tools::R_user_dir("biowR", "cache"),
  cwa_root,
  results_root,
  json_args = "",
  f0,
  f1,
  ncore,
  ht = FALSE
)

write_stage2_5_swarmfile(
  script_dir,
  output_dir,
  arguments_path,
  f0,
  f1,
  n_core,
  ht = FALSE
)

write_stage1_R_script(scriptDir = tools::R_user_dir("biowR", "cache"))

write_stage2_5_R_script(scriptDir = tools::R_user_dir("biowR"))
}
\arguments{
\item{scriptDir}{the directory where the R script is written}

\item{cwa_root}{The directory where the accelerometer files are stored.}

\item{results_root}{The root for the results directory.}

\item{json_args}{An optional json file with parameters for GGIR}

\item{f0}{The start index}

\item{f1}{The end index.  The file with index f1 is NOT run.}

\item{ncore}{The number of jobs you would like to swarm at once.}

\item{ht}{If you want to use hyper threading, set ht to TRUE.}
}
\value{
invisibly returns the name of the swarmfile

the name of the script file (invisibly)
}
\description{
Pay close attention to the parameters passed in. Currently, the code assumes that
each job takes 90 minutes to run.  This is to ensure that enough time is allocate to
run the job.

This is called from write_stage1_swarmfile or write_stage2_5_swarmfile
}
\details{
Because CRAN policy does not allow the scripts to be placed in the user's home directory without
permission, the default location of the scripts is in tools::R_user_dir("biowR"), which on biowulf
is ~/share/R/biowR.  I would suggest setting the scriptDir argument to something useful like "~".
}
\seealso{
\link{write_stage1_swarmfile}
}
