% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggir_swarm.R
\name{write_R_script}
\alias{write_R_script}
\alias{write_stage1_R_script}
\alias{write_stage2_5_R_script}
\title{Write the R script for running GGIR}
\usage{
write_stage1_R_script(scriptDir = tools::R_user_dir("biowR"))

write_stage2_5_R_script(scriptDir = tools::R_user_dir("biowR"))
}
\arguments{
\item{scriptDir}{the directory where the R script is written}
}
\value{
the name of the script file (invisibly)
}
\description{
This is called from write_stage1_swarmfile or write_stage2_5_swarmfile
}
\seealso{
\link{write_stage1_swarmfile}
}
